Use a separate `SqliteConnection` instance for each operation
While we already use separate `SqliteCommand` instances to prevent
cross-thread usage, there are no guarantees that the underlying
`SqliteConnection` instance is safe to be used across threads even with
separate `SqliteCommand` objects.

We now maintain a pool of open connections to the database alongside the
`SqliteCommand` pool and each time an `SqliteCommand` is requested, we
also rent a `SqliteConnection` and associate it with the `SqliteCommand`
before passing that to the caller.