Fix flaky SqliteCacheTests.BasicTests
This test wasn't safe to test in parallel, due to the hard-coded
database name/path.

In addition to using `[TestCleanup]` to clear the database from disk
after each test is run, we're also using an instance-specific
`Configuration.CachePath` that's created on class init, so it's now safe
to run the same test but for different frameworks in parallel (each gets
its own path).