Open newly created connections and log all open/close calls
I'm still getting random/racy exceptions in the `Dispose()` routine for
the bulk tests, so trace calls were added on the opening and closing of
all connections to the database to make sure they cancel out (and they
do).

This is likely caused by another process on the PC (antivirus? blergh!)
so it seems like everything is OK.