Implement `Cache.Clear()` without going through the command pool
Implement `SqliteCache.Clear()` without going through the command pool, using
the new `DbCommand.Use()` override that borrows a connection w/out a command,
then creating the command and using it locally.

Assuming that calls to `SqliteCache.Clear()` are not typically going to be in
the hot path, this will reduce the memory usage by not leaving cached instances
of the `DbCommand` around after the call.