Make compatible with `SQLITE_DSQ=0`
Fix SQL query errors when running against an SQLite build compiled with
`-DSQLITE_DSQ=0`, which makes SQLite behave like PostgreSQL and expect that
double quotes are only used to denote column name identifiers and never for
string literals.

So far as I can see, the only place we ever used double-quoted strings was in
initializing or querying the `meta` table; the "regular" queries mapped to the
`Operations` enum are all quote-free.

Closes #12.