Upgrade concrete nuget dependencies
Upgrade the concrete dependencies (i.e. excluding packages where we only use the
interface) to the latest version. The most important of these to note is the
upgrade of the SQLite dependency from `SQLitePCLRaw.bundle_green` 2.0.7 under
.NET Core and .NET 5+ to 2.1.4 (and from `SQLitePCLRaw.bundle_e_sqlite3` 2.0.7
to 2.1.4 under .NET Framework 4.6.x).

This is likely to fix the issues reported in #14 but we will need independent
confirmation of that as no repro steps were provided.