Add GitHub actions CI
Test under Linux, Windows, and macOS.

Configure actions/setup-dotnet to correctly test multiple .NET versions

We need to use a temporary global.json that contains the version of the .NET SDK
we want to run the tests under, as the default behavior is to run under the
greatest version installed regardless of what the project says.

Explicitly set framework for `dotnet` invocations

It seems `TargetFrameworks` is conflicting with our `global.json` CI step; we
are locking the SDK to a specific version but the csproj sometimes specifies a
higher target framework.

Use msbuild `-p` argument to specify TargetFrameworks override under CI

Revert CI build logs to "Normal" level

We've verified global.json is working and the correct version of the SDK is
invoked.